# Use obj is Array instead of obj.GetType().IsArray